requirements.txt
# Chat&Talk GPT - Python Dependencies
# FastAPI Backend

# Web Server
fastapi>=0.104.0
uvicorn>=0.24.0

# HTTP Requests
requests>=2.31.0
aiohttp>=3.9.0

# LLM APIs (Free)
groq>=0.4.0

# Text-to-Speech
edge-tts>=6.1.0
gTTS>=2.4.0
pyttsx3>=2.90
elevenlabs>=1.10.0

# Speech Recognition
SpeechRecognition>=3.10.0
whisper>=1.1.0

# Email Notifications
python-dotenv>=1.0.0

# Google Sheets Integration
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0

# Excel Export
pandas>=2.0.0
openpyxl>=3.0.0

# Research & Search Tools
googlesearch-python>=1.2.0
wikipedia-api>=0.6.0
duckduckgo-search>=4.4.0

# AI Provider APIs (for aggregator)
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.5.0
cohere>=5.0.0
mistralai>=1.0.0

# Authentication
PyJWT>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.0

# Optional: Ollama (local LLM)
# ollama>=0.1.0

# Mathematics and Problem Solving
sympy>=1.12.0
